AI/requirements.txt
pymongo==4.10.1
openai==1.45.0
python-dotenv==1.0.1
spacy==3.7.5
nltk==3.9.1
fastapi==0.115.0
uvicorn==0.30.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pandas==2.2.2
faiss-cpu==1.8.0.post1
pytest==8.3.2
flask==2.2.2
flask-cors==3.0.10
pymongo==4.10.1
geopy==2.2.0
requests==2.28.1

